Restructured PPMS plugin and tests Separated schemas for ETO and ACT
