+++ pyproject.toml
@@ -134,8 +134,10 @@ where = ["src"]
 general_schema = "nomad_measurements:schema"
 xrd_schema = "nomad_measurements.xrd:schema"
 xrd_parser = "nomad_measurements.xrd:parser"
-ppms_schema = "nomad_measurements.ppms:ppms_schema"
-ppms_data_parser = "nomad_measurements.ppms:ppms_data_parser"
-ppms_sequence_parser = "nomad_measurements.ppms:ppms_sequence_parser"
+ppms_eto_schema = "nomad_measurements.ppms:eto_schema"
+ppms_act_schema = "nomad_measurements.ppms:act_schema"
+ppms_eto_parser = "nomad_measurements.ppms:eto_parser"
+ppms_act_parser = "nomad_measurements.ppms:act_parser"
+ppms_sequence_parser = "nomad_measurements.ppms:sequence_parser"
 
 [tool.setuptools_scm]
